+background, bld ok
git-svn-id: https://subversion.assembla.com/svn/oddfry@9 572409ff-6c78-402f-b79f-49548ff2e957

+++ Code/OddFry/src/com/oddfry/graphics/Animation.java
@@ -104,7 +104,7 @@ public class Animation {
 	private int tab_[];
 	private float period_;
 	private int currentFrame_; 			//we print grid[tab_[currentFrame_]]
-	private float timeOnLastDraw_; 
+	private long timeOnLastDraw_; 
 	private float offsetCurrentFrame_; 	//will be added to currentFrame_ when >= 1
 	private boolean cycleEnded_ = false;
